+++ docs/conf.py
@@ -23,9 +23,9 @@ author = "NVIDIA"
 # built documents.
 #
 # The short X.Y version.
-version = "0.20"
+version = "21.12"
 # The full version, including alpha/beta/rc tags.
-release = "0.20.0"
+release = "21.12.00"
 
 
 # -- General configuration ---------------------------------------------------