+++ rapids-cmake/cpm/init.cmake
@@ -32,8 +32,8 @@ The CPM module will be downloaded based on the state of :cmake:variable:`CPM_SOU
 same download of CPM. If those variables aren't set the file will be cached
 in the build tree of the calling project
 
-``OVERRIDE``
 .. versionadded:: v21.10.00
+  ``OVERRIDE``
   Override the `CPM` preset package information for the project. The user provided
   json file must follow the `versions.json` format, which is :ref:`documented here<cpm_version_format>`.
 
@@ -47,10 +47,11 @@ in the build tree of the calling project
 function(rapids_cpm_init)
   list(APPEND CMAKE_MESSAGE_CONTEXT "rapids.cpm.init")
 
-  set(options)
-  set(one_value OVERRIDE)
-  set(multi_value)
-  cmake_parse_arguments(RAPIDS "${options}" "${one_value}" "${multi_value}" ${ARGN})
+  set(_rapids_options)
+  set(_rapids_one_value OVERRIDE)
+  set(_rapids_multi_value)
+  cmake_parse_arguments(RAPIDS "${_rapids_options}" "${_rapids_one_value}" "${_rapids_multi_value}"
+                        ${ARGN})
 
   include("${rapids-cmake-dir}/cpm/detail/load_preset_versions.cmake")
   rapids_cpm_load_preset_versions()
@@ -63,4 +64,9 @@ function(rapids_cpm_init)
   include("${rapids-cmake-dir}/cpm/detail/download.cmake")
   rapids_cpm_download()
 
+  # Propagate up any modified local variables that CPM has changed.
+  #
+  # Push up the modified CMAKE_MODULE_PATh to allow `find_package` calls to find packages that CPM
+  # already added.
+  set(CMAKE_MODULE_PATH "${CMAKE_MODULE_PATH}" PARENT_SCOPE)
 endfunction()